[procmgr] Split e2e tests into modules by use case (#55711)
### What does this PR do?

Splits the monolithic `tests/e2e.rs` into `tests/e2e/*.rs` modules
grouped by use case (daemon, fixtures, process control, list, reload,
describe, restart/burst, config/status, create, env files, CLI
contracts, DDOT). Updates Cargo and Bazel crate roots to
`tests/e2e/main.rs`.

Move-only: no test logic changes. DDOT tests stay in `ddot.rs` for now.

### Motivation

The TestEnv stack (#55627) put 91 e2e tests in one file. Smaller modules
make reviews easier without changing behavior.

### Describe how you validated your changes

- `cargo fmt --all`
- `cargo clippy --features test-helpers --tests -- -D warnings`
- `cargo test --features test-helpers --test e2e`
- `bazel test //pkg/procmgr/rust:e2e_test`

### Additional Notes

Stacks on `jose/procmgr-testenv` (#55627). Follow-ups: DDOT to new-e2e,
daemon shutdown charter, redundant CLI test cleanup.
