--- pkg/procmgr/rust/tests/e2e.rs
@@ -1,1919 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2026-present Datadog, Inc.
-
-mod helpers;
-
-use dd_procmgrd::test_helpers;
-use helpers::{
-    DescribeExpect, ProcessExpect, ReloadExpect, StatusProcessesCount, TestEnv, kill_pid_force,
-    pid_is_alive, wait_for_pid_gone, write_config,
-};
-use std::collections::BTreeMap;
-use std::time::Duration;
-
-#[test]
-fn daemon_starts_ready() {
-    let env = TestEnv::new();
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_version_not_empty();
-    status.assert_processes_count(StatusProcessesCount::zeros());
-}
-
-#[test]
-fn status_fails_without_daemon() {
-    let env = TestEnv::new();
-    env.status()
-        .expect_err("expected status to fail without daemon");
-}
-
-#[test]
-fn status_ready_without_config_dir() {
-    let env = TestEnv::with_missing_config_dir();
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount::zeros());
-}
-
-#[test]
-fn sleeper_fixture_auto_starts() {
-    let env = TestEnv::new().with_process("sleeper");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        running: Some(1),
-        created: Some(0),
-        ..Default::default()
-    });
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper", ProcessExpect::Running);
-}
-
-#[test]
-fn sleeper_fixture_no_auto_start() {
-    let env = TestEnv::new().with_process("sleeper_idle");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        created: Some(1),
-        running: Some(0),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Created);
-}
-
-#[test]
-fn invalid_syntax_fixture_skipped() {
-    let env = TestEnv::new()
-        .with_process("sleeper")
-        .with_process("invalid_syntax");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        running: Some(1),
-        created: Some(0),
-        ..Default::default()
-    });
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper", ProcessExpect::Running);
-    list.assert_absent("invalid_syntax");
-    procmgr.assert_config_skip_logged("invalid_syntax");
-}
-
-#[test]
-fn missing_binary_fixture_fails_to_spawn() {
-    let env = TestEnv::new().with_process("missing_binary");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        failed: Some(1),
-        running: Some(0),
-        created: Some(0),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("missing_binary", ProcessExpect::Failed);
-}
-
-#[test]
-fn exit_ok_fixture_exits_cleanly() {
-    let env = TestEnv::new().with_process("exit_ok");
-    let procmgr = env.start();
-    procmgr.assert_process_state_within("exit_ok", ProcessExpect::Exited);
-    let status = procmgr.require_status();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        exited: Some(1),
-        running: Some(0),
-        failed: Some(0),
-        ..Default::default()
-    });
-}
-
-#[test]
-fn exit_fail_fixture_exits_with_failure() {
-    let env = TestEnv::new().with_process("exit_fail");
-    let procmgr = env.start();
-    procmgr.assert_process_state_within("exit_fail", ProcessExpect::Failed);
-    let status = procmgr.require_status();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        failed: Some(1),
-        running: Some(0),
-        exited: Some(0),
-        ..Default::default()
-    });
-}
-
-#[test]
-fn condition_blocked_fixture_stays_created() {
-    let env = TestEnv::new().with_process("condition_blocked");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        created: Some(1),
-        running: Some(0),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("condition_blocked", ProcessExpect::Created);
-    procmgr.assert_condition_path_not_met_logged(
-        "condition_blocked",
-        "/nonexistent/path/procmgr-condition-test",
-    );
-}
-
-#[test]
-fn start_process_transitions_created_to_running() {
-    let env = TestEnv::new().with_process("sleeper_idle");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        created: Some(1),
-        running: Some(0),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Created);
-
-    procmgr.assert_start_process("sleeper_idle");
-
-    let status = procmgr.require_status();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        running: Some(1),
-        created: Some(0),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Running);
-}
-
-#[test]
-fn stop_process_transitions_running_to_stopped() {
-    let env = TestEnv::new().with_process("sleeper");
-    let procmgr = env.start();
-    let status = procmgr.require_status();
-    status.assert_ready();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        running: Some(1),
-        stopped: Some(0),
-        created: Some(0),
-        ..Default::default()
-    });
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-
-    procmgr.assert_stop_process("sleeper");
-
-    let status = procmgr.require_status();
-    status.assert_processes_count(StatusProcessesCount {
-        total: Some(1),
-        stopped: Some(1),
-        running: Some(0),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper", ProcessExpect::Stopped);
-}
-
-#[test]
-fn stop_process_kills_child() {
-    let env = TestEnv::new().with_process("sleeper");
-    let procmgr = env.start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-
-    let snapshot = procmgr
-        .process("sleeper")
-        .expect("sleeper should be listed");
-    let pid = snapshot.pid as u32;
-    assert!(pid > 0, "sleeper should have a PID, got {snapshot:?}");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive before stop");
-
-    procmgr.assert_stop_process("sleeper");
-
-    assert!(
-        wait_for_pid_gone(pid, Duration::from_secs(5)),
-        "child PID {pid} should be gone after stop"
-    );
-}
-
-#[test]
-fn list_empty_when_no_processes() {
-    let procmgr = TestEnv::new().start();
-    procmgr.require_status().assert_ready();
-    procmgr.require_list().assert_empty();
-}
-
-#[test]
-fn list_shows_running_and_created_mix() {
-    let env = TestEnv::new()
-        .with_process("sleeper")
-        .with_process("sleeper_idle");
-    let procmgr = env.start();
-    procmgr.require_status().assert_ready();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let list = procmgr.require_list();
-    list.assert_len(2);
-    list.assert_process_state("sleeper", ProcessExpect::Running);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Created);
-}
-
-#[test]
-fn list_shows_exited_with_last_exit_code() {
-    let env = TestEnv::new()
-        .with_process("exit_ok")
-        .with_process("exit_fail");
-    let procmgr = env.start();
-    procmgr.assert_process_state_within("exit_ok", ProcessExpect::Exited);
-    procmgr.assert_process_state_within("exit_fail", ProcessExpect::Failed);
-    let list = procmgr.require_list();
-    list.assert_len(2);
-    list.assert_last_exit_code("exit_ok", 0);
-    list.assert_last_exit_code("exit_fail", 1);
-}
-
-#[test]
-fn reload_unchanged_leaves_running_process() {
-    let procmgr = TestEnv::new().with_process("sleeper").start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    procmgr.assert_reload_matches(ReloadExpect {
-        unchanged: Some(vec!["sleeper".into()]),
-        added: Some(vec![]),
-        removed: Some(vec![]),
-        modified: Some(vec![]),
-        preserve_running_pids: vec!["sleeper".into()],
-    });
-}
-
-#[test]
-fn reload_adds_fixture_to_catalog() {
-    let procmgr = TestEnv::new().start();
-    procmgr.require_list().assert_empty();
-
-    procmgr.install_fixture("sleeper");
-    procmgr.assert_reload_matches(ReloadExpect {
-        added: Some(vec!["sleeper".into()]),
-        ..Default::default()
-    });
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper", ProcessExpect::Running);
-}
-
-#[test]
-fn reload_removes_deleted_yaml_from_catalog() {
-    let procmgr = TestEnv::new()
-        .with_process("sleeper")
-        .with_process("sleeper_idle")
-        .start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let list = procmgr.require_list();
-    list.assert_len(2);
-    list.assert_process_state("sleeper", ProcessExpect::Running);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Created);
-
-    procmgr.remove_process_yaml("sleeper_idle");
-    procmgr.assert_reload_matches(ReloadExpect {
-        removed: Some(vec!["sleeper_idle".into()]),
-        unchanged: Some(vec!["sleeper".into()]),
-        added: Some(vec![]),
-        modified: Some(vec![]),
-        preserve_running_pids: vec!["sleeper".into()],
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_absent("sleeper_idle");
-}
-
-#[test]
-fn reload_remove_only_empties_catalog() {
-    let procmgr = TestEnv::new().with_process("sleeper").start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    let old_pid = procmgr.process("sleeper").expect("sleeper").pid;
-
-    procmgr.remove_process_yaml("sleeper");
-    procmgr.assert_reload_matches(ReloadExpect {
-        removed: Some(vec!["sleeper".into()]),
-        added: Some(vec![]),
-        modified: Some(vec![]),
-        unchanged: Some(vec![]),
-        ..Default::default()
-    });
-
-    procmgr.assert_pid_gone(old_pid);
-    procmgr.require_list().assert_empty();
-}
-
-#[test]
-fn reload_adds_no_auto_start_stays_created() {
-    let procmgr = TestEnv::new().start();
-    procmgr.require_list().assert_empty();
-
-    procmgr.install_fixture("sleeper_idle");
-    procmgr.assert_reload_matches(ReloadExpect {
-        added: Some(vec!["sleeper_idle".into()]),
-        ..Default::default()
-    });
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Created);
-}
-
-#[test]
-fn reload_modify_respawns_with_new_pid() {
-    let procmgr = TestEnv::new().with_process("sleeper").start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let count_before = procmgr.require_list().len();
-    let old_pid = procmgr.process("sleeper").expect("sleeper").pid;
-
-    procmgr.overwrite_with_fixture("sleeper", "sleeper_long");
-
-    procmgr.assert_reload_matches(ReloadExpect {
-        modified: Some(vec!["sleeper".into()]),
-        added: Some(vec![]),
-        removed: Some(vec![]),
-        unchanged: Some(vec![]),
-        ..Default::default()
-    });
-
-    let count_after = procmgr.require_list().len();
-    assert_eq!(
-        count_before, count_after,
-        "modify reload should not change catalog size"
-    );
-    procmgr.assert_process_pid_changed("sleeper", old_pid);
-}
-
-#[test]
-fn reload_remove_and_add_swaps_catalog() {
-    let procmgr = TestEnv::new().with_process("sleeper").start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-    let old_pid = procmgr.process("sleeper").expect("sleeper").pid;
-
-    procmgr.remove_process_yaml("sleeper");
-    procmgr.install_fixture("sleeper_idle");
-
-    procmgr.assert_reload_matches(ReloadExpect {
-        removed: Some(vec!["sleeper".into()]),
-        added: Some(vec!["sleeper_idle".into()]),
-        modified: Some(vec![]),
-        unchanged: Some(vec![]),
-        ..Default::default()
-    });
-
-    procmgr.assert_pid_gone(old_pid);
-    let list = procmgr.require_list();
-    list.assert_len(1);
-    list.assert_process_state("sleeper_idle", ProcessExpect::Created);
-}
-
-#[test]
-fn describe_running_process_matches_fixture() {
-    let procmgr = TestEnv::new().with_process("sleeper").start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-
-    let list_snap = procmgr.process("sleeper").expect("sleeper");
-    procmgr.assert_describe_matches(
-        "sleeper",
-        DescribeExpect {
-            name: Some("sleeper".into()),
-            state: Some("Running".into()),
-            command: Some(list_snap.command),
-            args: Some(list_snap.args),
-            has_uuid: Some(true),
-            pid_alive: Some(true),
-            ..Default::default()
-        },
-    );
-}
-
-#[test]
-fn describe_resolves_uuid_prefix() {
-    let procmgr = TestEnv::new().with_process("sleeper").start();
-    procmgr
-        .wait_for_process_running("sleeper")
-        .expect("expected sleeper running");
-
-    let uuid = procmgr.process("sleeper").expect("sleeper").uuid.clone();
-    let prefix = uuid[..8].to_string();
-
-    for id in [&prefix, uuid.as_str()] {
-        procmgr.assert_describe_matches(
-            id,
-            DescribeExpect {
-                name: Some("sleeper".into()),
-                uuid: Some(uuid.clone()),
-                ..Default::default()
-            },
-        );
-    }
-}
-
-#[test]
-fn describe_shows_static_config_fields() {
-    let procmgr = TestEnv::new().with_process("full").start();
-    procmgr
-        .wait_for_process_running("full")
-        .expect("expected full running");
-
-    let root = procmgr.env_root().display().to_string();
-    procmgr.assert_describe_matches(
-        "full",
-        DescribeExpect {
-            name: Some("full".into()),
-            state: Some("Running".into()),
-            description: Some("a test process".into()),
-            working_dir: Some(root),
-            restart_policy: Some("always".into()),
-            auto_start: Some(true),
-            env_contains: Some(BTreeMap::from([("MY_VAR".into(), "hello".into())])),
-            after: Some(vec!["other".into()]),
-            has_stdout_path: Some(true),
-            has_stderr_path: Some(true),
-            ..Default::default()
-        },
-    );
-}
-
-#[test]
-fn describe_after_exit_shows_last_exit() {
-    let procmgr = TestEnv::new().with_process("exit_fail").start();
-    procmgr.assert_process_state_within("exit_fail", ProcessExpect::Failed);
-
-    procmgr.assert_describe_matches(
-        "exit_fail",
-        DescribeExpect {
-            name: Some("exit_fail".into()),
-            state: Some("Failed".into()),
-            pid: Some(0),
-            last_exit_code: Some(Some(1)),
-            ..Default::default()
-        },
-    );
-}
-
-#[test]
-fn restart_always_increments_count_in_list() {
-    let procmgr = TestEnv::new().with_process("crasher").start();
-    procmgr.assert_restart_count_at_least("crasher", 2);
-    procmgr.require_list().assert_len(1);
-}
-
-#[test]
-fn describe_shows_restart_count_after_always_policy() {
-    let procmgr = TestEnv::new().with_process("crasher").start();
-    procmgr.assert_restart_count_at_least("crasher", 2);
-    procmgr.assert_describe_matches(
-        "crasher",
-        DescribeExpect {
-            name: Some("crasher".into()),
-            restart_count_at_least: Some(2),
-            ..Default::default()
-        },
-    );
-}
-
-#[test]
-fn restart_on_failure_leaves_clean_exit_terminal() {
-    let procmgr = TestEnv::new().with_process("exit_ok_on_failure").start();
-    procmgr.assert_process_state_within("exit_ok_on_failure", ProcessExpect::Exited);
-    assert_eq!(
-        procmgr
-            .process("exit_ok_on_failure")
-            .expect("snap")
-            .restart_count,
-        0
-    );
-}
-
-#[test]
-fn restart_on_success_respawns_clean_exit() {
-    let procmgr = TestEnv::new().with_process("exit_ok_on_success").start();
-    procmgr.assert_restart_count_at_least("exit_ok_on_success", 2);
-}
-
-#[test]
-fn restart_on_success_leaves_failed_exit_terminal() {
-    let procmgr = TestEnv::new().with_process("exit_fail_on_success").start();
-    procmgr.assert_process_state_within("exit_fail_on_success", ProcessExpect::Failed);
-    assert_eq!(
-        procmgr
-            .process("exit_fail_on_success")
-            .expect("snap")
-            .restart_count,
-        0
-    );
-}
-
-#[test]
-fn restart_on_failure_respawns_failed_exit() {
-    let procmgr = TestEnv::new().with_process("exit_fail_on_failure").start();
-    procmgr.assert_restart_count_at_least("exit_fail_on_failure", 2);
-}
-
-#[test]
-fn burst_limit_stops_with_failed_state() {
-    let procmgr = TestEnv::new().with_process("burst_limited").start();
-    procmgr
-        .wait_for_restart_count_terminal(
-            "burst_limited",
-            4,
-            ProcessExpect::Failed,
-            Duration::from_secs(3),
-            Duration::from_secs(30),
-        )
-        .unwrap_or_else(|e| panic!("expected burst_limited to hit restart burst limit: {e}"));
-}
-
-#[test]
-fn burst_interval_allows_continued_restarts() {
-    let procmgr = TestEnv::new().with_process("burst_spaced").start();
-    procmgr
-        .wait_for_restart_count_at_least("burst_spaced", 5, Duration::from_secs(10))
-        .unwrap_or_else(|e| panic!("expected spaced restarts to continue: {e}"));
-}
-
-#[test]
-fn test_cli_config_basic() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    let config_dir = env.config_dir().display().to_string();
-
-    env.cli_config()
-        .assert_success()
-        .assert_field("Source", "yaml")
-        .assert_field("Location", &config_dir)
-        .assert_field("Loaded Processes", "1")
-        .assert_field("Runtime Processes", "0");
-}
-
-#[test]
-fn test_cli_config_json() {
-    let env = TestEnv::new()
-        .with_config("svc-a", test_helpers::sleep_config_yaml())
-        .with_config("svc-b", test_helpers::sleep_config_yaml())
-        .start();
-
-    let config_dir = env.config_dir().display().to_string();
-
-    let out = env.cli_config_json();
-    out.assert_success();
-    let json = out.stdout_json();
-
-    assert_eq!(json["source"], "yaml");
-    assert_eq!(json["location"], config_dir.as_str());
-    assert_eq!(json["loaded_processes"], 2);
-    assert_eq!(json["runtime_processes"], 0);
-}
-
-#[test]
-fn test_cli_status_basic() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Ready", "true")
-        .assert_has_field("Version")
-        .assert_has_field("Uptime")
-        .assert_field("Total Processes", "1")
-        .assert_field("Running", "1")
-        .assert_field("Stopped", "0")
-        .assert_field("Created", "0")
-        .assert_field("Failed", "0")
-        .assert_field("Exited", "0");
-}
-
-#[test]
-fn test_cli_status_counts() {
-    let env = TestEnv::new()
-        .with_config("runner-a", test_helpers::sleep_config_yaml())
-        .with_config("runner-b", test_helpers::sleep_config_yaml())
-        .with_config(
-            "idle",
-            &test_helpers::sleep_config_with("auto_start: false\n"),
-        )
-        .start();
-
-    env.daemon().wait_for_log_default("[runner-a] spawned");
-    env.daemon().wait_for_log_default("[runner-b] spawned");
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Total Processes", "3")
-        .assert_field("Running", "2")
-        .assert_field("Created", "1")
-        .assert_field("Stopped", "0")
-        .assert_field("Failed", "0")
-        .assert_field("Exited", "0");
-}
-
-#[test]
-fn test_cli_status_json() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .with_config(
-            "idle",
-            &test_helpers::sleep_config_with("auto_start: false\n"),
-        )
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let out = env.cli_status_json();
-    out.assert_success();
-    let json = out.stdout_json();
-
-    assert_eq!(json["ready"], true);
-    assert!(!json["version"].as_str().unwrap_or("").is_empty());
-    assert!(json["uptime_seconds"].as_u64().is_some());
-    assert_eq!(json["total_processes"], 2);
-    assert_eq!(json["running_processes"], 1);
-    assert_eq!(json["created_processes"], 1);
-    assert_eq!(json["stopped_processes"], 0);
-    assert_eq!(json["failed_processes"], 0);
-    assert_eq!(json["exited_processes"], 0);
-    assert_eq!(json["starting_processes"], 0);
-    assert_eq!(json["stopping_processes"], 0);
-}
-
-#[test]
-fn test_cli_status_after_stop() {
-    let env = TestEnv::new()
-        .with_config("svc-a", test_helpers::sleep_config_yaml())
-        .with_config("svc-b", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[svc-a] spawned");
-    env.daemon().wait_for_log_default("[svc-b] spawned");
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Running", "2")
-        .assert_field("Stopped", "0");
-
-    env.cli_stop("svc-a").assert_success();
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Total Processes", "2")
-        .assert_field("Running", "1")
-        .assert_field("Stopped", "1");
-
-    let json = env.cli_status_json().stdout_json();
-    assert_eq!(json["running_processes"], 1);
-    assert_eq!(json["stopped_processes"], 1);
-}
-
-#[test]
-fn test_cli_status_mixed_states() {
-    let env = TestEnv::new()
-        .with_config("runner", test_helpers::sleep_config_yaml())
-        .with_config("bad", "command: /nonexistent/binary\n")
-        .with_config("quick", test_helpers::true_config_yaml())
-        .with_config(
-            "idle",
-            &test_helpers::sleep_config_with("auto_start: false\n"),
-        )
-        .start();
-
-    env.daemon().wait_for_log_default("[runner] spawned");
-    env.daemon().wait_for_log_default("[bad] failed to spawn");
-    env.daemon().wait_for_log_default("[quick] exited with");
-
-    let out = env.cli_status();
-    out.assert_success()
-        .assert_field("Total Processes", "4")
-        .assert_field("Running", "1")
-        .assert_field("Created", "1")
-        .assert_field("Failed", "1")
-        .assert_field("Exited", "1")
-        .assert_field("Stopped", "0");
-
-    let json_out = env.cli_status_json();
-    json_out.assert_success();
-    let json = json_out.stdout_json();
-    assert_eq!(json["total_processes"], 4);
-    assert_eq!(json["running_processes"], 1);
-    assert_eq!(json["created_processes"], 1);
-    assert_eq!(json["failed_processes"], 1);
-    assert_eq!(json["exited_processes"], 1);
-    assert_eq!(json["stopped_processes"], 0);
-}
-
-#[test]
-fn test_cli_config_with_runtime_processes() {
-    let env = TestEnv::new()
-        .with_config("loaded", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[loaded] spawned");
-
-    env.cli_config()
-        .assert_success()
-        .assert_field("Loaded Processes", "1")
-        .assert_field("Runtime Processes", "0");
-
-    env.create_sleep("dynamic", &[]).assert_success();
-
-    env.cli_config()
-        .assert_success()
-        .assert_field("Loaded Processes", "1")
-        .assert_field("Runtime Processes", "1");
-}
-
-#[test]
-fn test_cli_list_terminal_table_fields() {
-    let procmgr = TestEnv::new()
-        .with_process("exit_ok")
-        .with_process("exit_fail")
-        .start();
-    procmgr.assert_process_state_within("exit_ok", ProcessExpect::Exited);
-    procmgr.assert_process_state_within("exit_fail", ProcessExpect::Failed);
-
-    let python = test_helpers::python_exe();
-    procmgr
-        .cli_list()
-        .assert_success()
-        .assert_table_row(
-            "exit_ok",
-            &[
-                ("STATE", "Exited"),
-                ("PID", "-"),
-                ("LAST EXIT", "exit 0"),
-                ("COMMAND", &python),
-            ],
-        )
-        .assert_table_row(
-            "exit_fail",
-            &[
-                ("STATE", "Failed"),
-                ("PID", "-"),
-                ("LAST EXIT", "exit 1"),
-                ("COMMAND", &python),
-            ],
-        )
-        .assert_table_row_count(2);
-}
-
-#[test]
-fn test_cli_list_json() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let out = env.cli_list_json();
-    out.assert_success();
-    let json = out.stdout_json();
-    let arr = json.as_array().expect("expected JSON array");
-    assert_eq!(arr.len(), 1);
-
-    let entry = &arr[0];
-    assert_eq!(entry["name"], "sleeper");
-    assert_eq!(entry["state"], "Running");
-    assert_eq!(entry["command"], test_helpers::sleep_cmd(300).0);
-    assert_eq!(entry["args"], test_helpers::sleep_args_json());
-    assert_eq!(entry["restart_count"], 0);
-    assert!(!entry["uuid"].as_str().unwrap_or("").is_empty());
-    assert!(entry["last_exit_code"].is_null());
-    assert!(entry["last_signal"].is_null());
-
-    let pid = entry["pid"].as_u64().expect("pid should be a number") as u32;
-    assert!(pid > 0, "running process should have a PID");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-}
-
-#[test]
-fn test_cli_list_json_empty() {
-    let env = TestEnv::new().start();
-
-    let out = env.cli_list_json();
-    out.assert_success();
-    let json = out.stdout_json();
-    let arr = json.as_array().expect("expected JSON array");
-    assert!(arr.is_empty(), "expected empty array, got {json}");
-}
-
-#[test]
-fn test_cli_describe_last_exit_text() {
-    let env = TestEnv::new().with_process("exit_fail").start();
-    env.assert_process_state_within("exit_fail", ProcessExpect::Failed);
-
-    env.cli_describe("exit_fail")
-        .assert_success()
-        .assert_field("Last Exit", "exit 1");
-}
-
-#[test]
-fn test_cli_describe_restarts_text() {
-    let env = TestEnv::new().with_process("crasher").start();
-    env.assert_restart_count_at_least("crasher", 2);
-
-    env.cli_describe("crasher")
-        .assert_success()
-        .assert_field("Name", "crasher")
-        .assert_field_at_least("Restarts", 2);
-}
-
-#[test]
-fn test_cli_describe_not_found() {
-    let env = TestEnv::new().start();
-
-    env.cli_describe("nonexistent")
-        .assert_failure()
-        .assert_stderr_contains("not found");
-}
-
-#[test]
-fn test_cli_describe_json() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let out = env.cli_describe_json("sleeper");
-    out.assert_success();
-    let json = out.stdout_json();
-
-    assert_eq!(json["name"], "sleeper");
-    assert_eq!(json["state"], "Running");
-    assert_eq!(json["command"], test_helpers::sleep_cmd(300).0);
-    assert_eq!(json["args"], test_helpers::sleep_args_json());
-    assert!(!json["uuid"].as_str().unwrap_or("").is_empty());
-
-    let pid = json["pid"].as_u64().expect("pid should be a number") as u32;
-    assert!(pid > 0);
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-}
-
-#[test]
-fn test_cli_start_by_uuid() {
-    let env = TestEnv::new()
-        .with_config(
-            "sleeper",
-            &test_helpers::sleep_config_with("auto_start: false\n"),
-        )
-        .start();
-
-    let list_json = env.cli_list_json().stdout_json();
-    let uuid = list_json[0]["uuid"]
-        .as_str()
-        .expect("uuid should be a string");
-    let prefix = &uuid[..8];
-
-    env.cli_start(prefix)
-        .assert_success()
-        .assert_field("State", "Running");
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let pid = env.cli_describe("sleeper").pid_from_field("PID");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-}
-
-#[test]
-fn test_cli_start_already_running() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    env.cli_start("sleeper")
-        .assert_failure()
-        .assert_stderr_contains("already");
-}
-
-#[test]
-fn test_cli_start_not_found() {
-    let env = TestEnv::new().start();
-
-    env.cli_start("nonexistent")
-        .assert_failure()
-        .assert_stderr_contains("not found");
-}
-
-#[test]
-fn test_cli_start_json() {
-    let env = TestEnv::new()
-        .with_config(
-            "sleeper",
-            &test_helpers::sleep_config_with("auto_start: false\n"),
-        )
-        .start();
-
-    let out = env.cli_start_json("sleeper");
-    out.assert_success();
-    let json = out.stdout_json();
-
-    assert!(!json["uuid"].as_str().unwrap_or("").is_empty());
-    assert_eq!(json["state"], "Running");
-
-    let pid = json["pid"].as_u64().expect("pid should be a number") as u32;
-    assert!(pid > 0, "started process should have a PID");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-}
-
-#[test]
-fn test_cli_stop_by_uuid() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let list_json = env.cli_list_json().stdout_json();
-    let uuid = list_json[0]["uuid"]
-        .as_str()
-        .expect("uuid should be a string");
-    let prefix = &uuid[..8];
-
-    env.cli_stop(prefix)
-        .assert_success()
-        .assert_field("State", "Stopped");
-}
-
-#[test]
-fn test_cli_stop_already_stopped() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    env.cli_stop("sleeper").assert_success();
-    env.daemon().wait_for_log_default("[sleeper] stopped");
-
-    env.cli_stop("sleeper")
-        .assert_failure()
-        .assert_stderr_contains("not running");
-}
-
-#[test]
-fn test_cli_stop_not_found() {
-    let env = TestEnv::new().start();
-
-    env.cli_stop("nonexistent")
-        .assert_failure()
-        .assert_stderr_contains("not found");
-}
-
-#[test]
-fn test_cli_stop_json() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let out = env.cli_stop_json("sleeper");
-    out.assert_success();
-    let json = out.stdout_json();
-
-    assert!(!json["uuid"].as_str().unwrap_or("").is_empty());
-    assert_eq!(json["state"], "Stopped");
-}
-
-#[test]
-fn test_cli_create_minimal() {
-    let env = TestEnv::new().start();
-
-    let out = env.create_sleep("foo", &[]);
-    out.assert_success().assert_has_field("UUID");
-
-    env.daemon().wait_for_log_default("[foo] spawned");
-
-    let pid = env.cli_describe("foo").pid_from_field("PID");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-}
-
-#[test]
-fn test_cli_create_with_auto_start() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("svc", &[]).assert_success();
-
-    env.daemon().wait_for_log_default("[svc] spawned");
-
-    env.cli_list()
-        .assert_success()
-        .assert_table_row("svc", &[("STATE", "Running")]);
-
-    let pid = env.cli_list().pid_from_table_row("svc");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-}
-
-#[test]
-fn test_cli_create_no_auto_start() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("manual", &["--no-auto-start"])
-        .assert_success();
-
-    env.cli_list()
-        .assert_success()
-        .assert_table_row("manual", &[("STATE", "Created"), ("PID", "-")]);
-}
-
-#[test]
-fn test_cli_create_with_all_options() {
-    let env = TestEnv::new()
-        .with_config("dep", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[dep] spawned");
-
-    let tmp = test_helpers::temp_dir_str();
-    env.create_sleep(
-        "full",
-        &[
-            "--env",
-            "KEY1=val1",
-            "--env",
-            "KEY2=val2",
-            "--working-dir",
-            &tmp,
-            "--restart-policy",
-            "always",
-            "--description",
-            "full test",
-            "--after",
-            "dep",
-        ],
-    )
-    .assert_success();
-
-    env.daemon().wait_for_log_default("[full] spawned");
-
-    let out = env.cli_describe("full");
-    out.assert_success()
-        .assert_field("Name", "full")
-        .assert_field("Command", test_helpers::sleep_cmd(300).0)
-        .assert_field("Working Dir", &tmp)
-        .assert_field("Restart Policy", "always")
-        .assert_field("Description", "full test");
-}
-
-#[test]
-fn test_cli_create_then_describe() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("svc", &["--no-auto-start"])
-        .assert_success();
-
-    let out = env.cli_describe("svc");
-    out.assert_success()
-        .assert_field("Name", "svc")
-        .assert_field("State", "Created")
-        .assert_field("Command", test_helpers::sleep_cmd(300).0)
-        .assert_field("Args", &test_helpers::sleep_args_display())
-        .assert_field("PID", "-")
-        .assert_has_field("UUID");
-}
-
-#[test]
-fn test_cli_create_duplicate_name() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("dup", &[]).assert_success();
-
-    env.daemon().wait_for_log_default("[dup] spawned");
-
-    env.create_sleep("dup", &[])
-        .assert_failure()
-        .assert_stderr_contains("already exists");
-}
-
-#[test]
-fn test_cli_create_empty_command() {
-    let env = TestEnv::new().start();
-
-    env.cli(&["create", "--name", "foo", "--command", ""])
-        .assert_failure()
-        .assert_stderr_contains("command must not be empty");
-}
-
-#[test]
-fn test_cli_create_invalid_name() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("bad name!", &[])
-        .assert_failure()
-        .assert_stderr_contains("name must only contain");
-}
-
-#[test]
-fn test_cli_create_json() {
-    let env = TestEnv::new().start();
-
-    let out = env.create_sleep("svc", &["--json"]);
-    out.assert_success();
-    let json = out.stdout_json();
-
-    assert_eq!(json["name"], "svc");
-    assert!(!json["uuid"].as_str().unwrap_or("").is_empty());
-}
-
-#[test]
-fn test_cli_create_env_vars() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep(
-        "env-svc",
-        &["--env", "FOO=bar", "--env", "BAZ=qux", "--no-auto-start"],
-    )
-    .assert_success();
-
-    let out = env.cli_describe_json("env-svc");
-    out.assert_success();
-    let json = out.stdout_json();
-
-    let env_map = json["env"].as_object().expect("env should be an object");
-    assert_eq!(env_map.get("FOO").and_then(|v| v.as_str()), Some("bar"));
-    assert_eq!(env_map.get("BAZ").and_then(|v| v.as_str()), Some("qux"));
-}
-
-#[test]
-fn test_cli_reload_text_modified() {
-    let env = TestEnv::new().with_process("sleeper").start();
-    env.overwrite_with_fixture("sleeper", "sleeper_long");
-
-    env.cli_reload()
-        .assert_success()
-        .assert_field("Modified", "sleeper");
-}
-
-#[test]
-fn test_cli_reload_text_added_removed() {
-    let env = TestEnv::new().with_process("sleeper").start();
-    env.remove_process_yaml("sleeper");
-    env.install_fixture("sleeper_idle");
-
-    env.cli_reload()
-        .assert_success()
-        .assert_field("Added", "sleeper_idle")
-        .assert_field("Removed", "sleeper");
-}
-
-#[test]
-fn test_cli_reload_json() {
-    let env = TestEnv::new()
-        .with_config("existing", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[existing] spawned");
-
-    write_config(env.config_dir(), "added", test_helpers::sleep_config_yaml());
-
-    let out = env.cli_reload_json();
-    out.assert_success();
-    let json = out.stdout_json();
-
-    let added = json["added"].as_array().expect("added should be an array");
-    assert!(
-        added.iter().any(|v| v.as_str() == Some("added")),
-        "expected 'added' in added array: {json}"
-    );
-
-    let unchanged = json["unchanged"]
-        .as_array()
-        .expect("unchanged should be an array");
-    assert!(
-        unchanged.iter().any(|v| v.as_str() == Some("existing")),
-        "expected 'existing' in unchanged array: {json}"
-    );
-
-    assert!(json["removed"].as_array().expect("array").is_empty());
-    assert!(json["modified"].as_array().expect("array").is_empty());
-}
-
-#[test]
-fn test_cli_full_lifecycle() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("svc", &["--no-auto-start"])
-        .assert_success();
-
-    env.cli_list()
-        .assert_success()
-        .assert_table_row("svc", &[("STATE", "Created"), ("PID", "-")]);
-
-    env.cli_start("svc").assert_success();
-    env.daemon().wait_for_log_default("[svc] spawned");
-
-    let out = env.cli_describe("svc");
-    out.assert_success().assert_field("State", "Running");
-    let pid = out.pid_from_field("PID");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive");
-
-    env.cli_stop("svc").assert_success();
-
-    assert!(
-        wait_for_pid_gone(pid, Duration::from_secs(5)),
-        "PID {pid} should be gone after stop"
-    );
-
-    env.cli_describe("svc")
-        .assert_success()
-        .assert_field("State", "Stopped")
-        .assert_field("PID", "-");
-}
-
-#[test]
-fn test_cli_create_stop_start_cycle() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("svc", &[]).assert_success();
-    env.daemon().wait_for_log_default("[svc] spawned");
-
-    env.cli_describe("svc")
-        .assert_success()
-        .assert_field("State", "Running");
-
-    env.cli_stop("svc").assert_success();
-    env.cli_describe("svc")
-        .assert_success()
-        .assert_field("State", "Stopped");
-
-    env.cli_start("svc").assert_success();
-    env.daemon()
-        .wait_for_log_count("[svc] spawned", 2, Duration::from_secs(10));
-    env.cli_describe("svc")
-        .assert_success()
-        .assert_field("State", "Running");
-
-    let pid = env.cli_describe("svc").pid_from_field("PID");
-    assert!(pid_is_alive(pid), "PID {pid} should be alive after restart");
-
-    env.cli_stop("svc").assert_success();
-    assert!(
-        wait_for_pid_gone(pid, Duration::from_secs(5)),
-        "PID {pid} should be gone after second stop"
-    );
-
-    env.cli_start("svc").assert_success();
-    env.daemon()
-        .wait_for_log_count("[svc] spawned", 3, Duration::from_secs(10));
-
-    let new_pid = env.cli_describe("svc").pid_from_field("PID");
-    assert!(pid_is_alive(new_pid), "PID {new_pid} should be alive");
-}
-
-#[test]
-fn test_cli_stop_start_then_kill_restarts_on_failure() {
-    let env = TestEnv::new()
-        .with_config(
-            "sleeper",
-            &test_helpers::sleep_config_with("restart: on-failure\n"),
-        )
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    env.cli_stop("sleeper").assert_success();
-    env.cli_start("sleeper").assert_success();
-    assert!(
-        env.daemon()
-            .wait_for_log_count("[sleeper] spawned", 2, Duration::from_secs(10)),
-        "start after stop should spawn a new child"
-    );
-
-    let pid = env.cli_describe("sleeper").pid_from_field("PID");
-    kill_pid_force(pid);
-
-    assert!(
-        env.daemon()
-            .wait_for_log_count("[sleeper] spawned", 3, Duration::from_secs(10)),
-        "on-failure should auto-restart after stop -> start -> external kill"
-    );
-
-    let json = env.cli_list_json().stdout_json();
-    assert_eq!(json[0]["state"], "Running");
-    assert!(
-        json[0]["restart_count"].as_u64().unwrap() >= 1,
-        "restart_count should reflect the crash restart"
-    );
-}
-
-#[test]
-fn test_cli_status_reflects_operations() {
-    let env = TestEnv::new()
-        .with_config("svc-a", test_helpers::sleep_config_yaml())
-        .with_config("svc-b", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[svc-a] spawned");
-    env.daemon().wait_for_log_default("[svc-b] spawned");
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Total Processes", "2")
-        .assert_field("Running", "2")
-        .assert_field("Stopped", "0")
-        .assert_field("Created", "0")
-        .assert_field("Failed", "0")
-        .assert_field("Exited", "0");
-
-    env.cli_stop("svc-a").assert_success();
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Total Processes", "2")
-        .assert_field("Running", "1")
-        .assert_field("Stopped", "1")
-        .assert_field("Created", "0")
-        .assert_field("Failed", "0")
-        .assert_field("Exited", "0");
-
-    env.create_sleep("svc-c", &[]).assert_success();
-    env.daemon().wait_for_log_default("[svc-c] spawned");
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Total Processes", "3")
-        .assert_field("Running", "2")
-        .assert_field("Stopped", "1")
-        .assert_field("Created", "0")
-        .assert_field("Failed", "0")
-        .assert_field("Exited", "0");
-}
-
-#[test]
-fn test_cli_create_with_dependencies() {
-    let env = TestEnv::new().start();
-
-    env.create_sleep("backend", &["--no-auto-start"])
-        .assert_success();
-
-    env.create_sleep("frontend", &["--after", "backend", "--no-auto-start"])
-        .assert_success();
-
-    env.cli_list()
-        .assert_success()
-        .assert_table_row_count(2)
-        .assert_table_row("backend", &[("STATE", "Created"), ("PID", "-")])
-        .assert_table_row("frontend", &[("STATE", "Created"), ("PID", "-")]);
-
-    env.cli_start("backend").assert_success();
-    env.daemon().wait_for_log_default("[backend] spawned");
-    env.cli_start("frontend").assert_success();
-    env.daemon().wait_for_log_default("[frontend] spawned");
-
-    let backend_pid = env.cli_list().pid_from_table_row("backend");
-    let frontend_pid = env.cli_list().pid_from_table_row("frontend");
-    assert!(
-        pid_is_alive(backend_pid),
-        "backend PID {backend_pid} should be alive"
-    );
-    assert!(
-        pid_is_alive(frontend_pid),
-        "frontend PID {frontend_pid} should be alive"
-    );
-
-    let out = env.cli_describe_json("frontend");
-    out.assert_success();
-    let json = out.stdout_json();
-    let after = json["after"].as_array().expect("after should be an array");
-    assert!(
-        after.iter().any(|v| v.as_str() == Some("backend")),
-        "frontend should depend on backend: {json}"
-    );
-}
-
-#[test]
-fn test_cli_create_nonexistent_dependency_ignored() {
-    let env = TestEnv::new().start();
-
-    let out = env.create_sleep("svc", &["--after", "does-not-exist"]);
-    out.assert_success();
-    out.assert_stderr_contains("not found, ignoring");
-
-    env.daemon().wait_for_log_default("[svc] spawned");
-
-    let pid = env.cli_list().pid_from_table_row("svc");
-    assert!(
-        pid_is_alive(pid),
-        "PID {pid} should be alive despite missing dep"
-    );
-
-    let out = env.cli_describe_json("svc");
-    out.assert_success();
-    let json = out.stdout_json();
-    let after = json["after"].as_array().expect("after should be an array");
-    assert!(
-        after.iter().any(|v| v.as_str() == Some("does-not-exist")),
-        "after should still list the nonexistent dep: {json}"
-    );
-}
-
-#[test]
-fn test_cli_create_nonexistent_dependency_json_warnings() {
-    let env = TestEnv::new().start();
-
-    let out = env.create_sleep("svc", &["--after", "ghost", "--json"]);
-    out.assert_success();
-    let json = out.stdout_json();
-    let warnings = json["warnings"]
-        .as_array()
-        .expect("warnings should be an array");
-    assert!(
-        warnings.iter().any(|w| {
-            let s = w.as_str().unwrap_or("");
-            s.contains("ghost") && s.contains("not found")
-        }),
-        "JSON warnings should mention missing dep: {json}"
-    );
-}
-
-#[test]
-fn test_cli_all_commands_json_parseable() {
-    let env = TestEnv::new()
-        .with_config("svc", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[svc] spawned");
-
-    let commands: Vec<&[&str]> = vec![
-        &["status", "--json"],
-        &["list", "--json"],
-        &["describe", "--json", "svc"],
-        &["config", "--json"],
-    ];
-    for args in &commands {
-        let out = env.cli(args);
-        out.assert_success();
-        out.stdout_json();
-    }
-
-    env.cli_stop_json("svc").assert_success().stdout_json();
-
-    env.cli_start_json("svc").assert_success().stdout_json();
-
-    env.create_sleep("dyn", &["--json", "--no-auto-start"])
-        .assert_success()
-        .stdout_json();
-
-    write_config(
-        env.config_dir(),
-        "extra",
-        &test_helpers::sleep_config_with("auto_start: false\n"),
-    );
-    env.cli_reload_json().assert_success().stdout_json();
-}
-
-#[test]
-fn test_cli_errors_on_stderr() {
-    let env = TestEnv::new().start();
-
-    let cases: Vec<(&[&str], &str)> = vec![
-        (&["describe", "nonexistent"], "not found"),
-        (&["start", "nonexistent"], "not found"),
-        (&["stop", "nonexistent"], "not found"),
-    ];
-    for (args, pattern) in &cases {
-        let out = env.cli(args);
-        out.assert_failure();
-        out.assert_stderr_contains(pattern);
-        assert!(
-            out.stdout.trim().is_empty(),
-            "stdout should be empty on error for {:?}, got: {}",
-            args,
-            out.stdout
-        );
-    }
-}
-
-#[test]
-fn test_cli_exit_codes() {
-    let env = TestEnv::new()
-        .with_config("svc", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[svc] spawned");
-
-    let success_cmds: Vec<&[&str]> =
-        vec![&["status"], &["list"], &["describe", "svc"], &["config"]];
-    for args in &success_cmds {
-        let out = env.cli(args);
-        assert_eq!(out.status.code(), Some(0), "expected exit 0 for {:?}", args);
-    }
-
-    let failure_cmds: Vec<&[&str]> = vec![
-        &["describe", "nonexistent"],
-        &["start", "nonexistent"],
-        &["stop", "nonexistent"],
-    ];
-    for args in &failure_cmds {
-        let out = env.cli(args);
-        assert_eq!(out.status.code(), Some(1), "expected exit 1 for {:?}", args);
-    }
-}
-
-#[test]
-fn test_cli_daemon_nonexistent_config_dir() {
-    let env = TestEnv::with_missing_config_dir();
-    let config_dir = env.config_dir().display().to_string();
-    let env = env.start();
-
-    env.daemon().wait_for_log_default(&format!(
-        "config directory {config_dir} does not exist, no processes to manage"
-    ));
-
-    env.cli_status()
-        .assert_success()
-        .assert_field("Ready", "true")
-        .assert_field("Total Processes", "0");
-
-    env.cli_list()
-        .assert_success()
-        .assert_stdout_contains("No processes");
-}
-
-#[test]
-fn test_cli_daemon_shutdown_stops_children() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let pid = env.cli_list().pid_from_table_row("sleeper");
-    assert!(pid_is_alive(pid), "child PID {pid} should be alive");
-
-    drop(env);
-
-    assert!(
-        wait_for_pid_gone(pid, Duration::from_secs(5)),
-        "child PID {pid} should be gone after daemon shutdown"
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn test_cli_daemon_shutdown_via_sigint() {
-    let env = TestEnv::new()
-        .with_config("sleeper", test_helpers::sleep_config_yaml())
-        .start();
-
-    env.daemon().wait_for_log_default("[sleeper] spawned");
-
-    let child_pid = env.cli_list().pid_from_table_row("sleeper");
-    let daemon_pid = env.daemon_pid();
-
-    assert!(pid_is_alive(child_pid), "child should be alive");
-    assert!(pid_is_alive(daemon_pid), "daemon should be alive");
-
-    env.daemon().send_signal(nix::sys::signal::Signal::SIGINT);
-    drop(env);
-
-    assert!(
-        !pid_is_alive(daemon_pid),
-        "daemon PID {daemon_pid} should be gone after SIGINT"
-    );
-    assert!(
-        !pid_is_alive(child_pid),
-        "child PID {child_pid} should be gone after SIGINT shutdown"
-    );
-}
-
-#[test]
-fn test_cli_condition_path_exists_not_met() {
-    let env = TestEnv::new()
-        .with_config(
-            "missing-bin",
-            "command: /nonexistent/binary\ncondition_path_exists: /nonexistent/binary\n",
-        )
-        .start();
-
-    env.daemon()
-        .wait_for_log_default("[missing-bin] condition_path_exists not met");
-
-    let json = env.cli_list_json().stdout_json();
-    assert_eq!(json[0]["name"], "missing-bin");
-    assert_eq!(json[0]["state"], "Created");
-    assert_eq!(json[0]["pid"], 0);
-}
-
-#[cfg(unix)]
-#[test]
-fn test_cli_environment_file_loading() {
-    let env = TestEnv::new();
-    let env_file = env.config_dir().join("test.env");
-    std::fs::write(&env_file, "MY_VAR=from_file\n").unwrap();
-
-    let env = env
-        .with_config(
-            "env-test",
-            &format!(
-                concat!(
-                    "command: /bin/sh\n",
-                    "args:\n",
-                    "  - '-c'\n",
-                    "  - 'exit $(test \"$MY_VAR\" = \"from_file\" && echo 0 || echo 1)'\n",
-                    "environment_file: {}\n",
-                    "restart: never\n",
-                ),
-                env_file.display()
-            ),
-        )
-        .start();
-
-    env.daemon().wait_for_log_default("[env-test] exited with");
-
-    let json = env.cli_list_json().stdout_json();
-    assert_eq!(json[0]["state"], "Exited");
-    assert_eq!(json[0]["last_exit_code"], 0);
-}
-
-#[cfg(unix)]
-#[test]
-fn test_cli_env_overrides_environment_file() {
-    let env = TestEnv::new();
-    let env_file = env.config_dir().join("test.env");
-    std::fs::write(&env_file, "MY_VAR=from_file\n").unwrap();
-
-    let env = env
-        .with_config(
-            "override-test",
-            &format!(
-                concat!(
-                    "command: /bin/sh\n",
-                    "args:\n",
-                    "  - '-c'\n",
-                    "  - 'exit $(test \"$MY_VAR\" = \"overridden\" && echo 0 || echo 1)'\n",
-                    "environment_file: {}\n",
-                    "env:\n",
-                    "  MY_VAR: overridden\n",
-                    "restart: never\n",
-                ),
-                env_file.display()
-            ),
-        )
-        .start();
-
-    env.daemon()
-        .wait_for_log_default("[override-test] exited with");
-
-    let json = env.cli_list_json().stdout_json();
-    assert_eq!(json[0]["state"], "Exited");
-    assert_eq!(json[0]["last_exit_code"], 0);
-}
-
-#[cfg(unix)]
-#[test]
-fn test_cli_child_does_not_inherit_parent_env() {
-    let env = TestEnv::new()
-        .with_config(
-            "clean-env",
-            concat!(
-                "command: /bin/sh\n",
-                "args:\n",
-                "  - '-c'\n",
-                "  - 'test -z \"$HOME\" && exit 0 || exit 1'\n",
-                "restart: never\n",
-            ),
-        )
-        .start();
-
-    env.daemon().wait_for_log_default("[clean-env] exited with");
-
-    let json = env.cli_list_json().stdout_json();
-    assert_eq!(json[0]["state"], "Exited");
-    assert_eq!(json[0]["last_exit_code"], 0);
-}
-
-#[test]
-fn test_cli_optional_environment_file_skipped_when_missing() {
-    let (sh, flag) = test_helpers::shell_cmd();
-    let env = TestEnv::new()
-        .with_config(
-            "opt-env",
-            &test_helpers::cmd_yaml(
-                sh,
-                &[flag.to_string(), "exit 0".to_string()],
-                "environment_file: -/nonexistent/env\nrestart: never\n",
-            ),
-        )
-        .start();
-
-    env.daemon()
-        .wait_for_log_default("optional environment file not found, skipping");
-    env.daemon().wait_for_log_default("[opt-env] exited with");
-
-    let json = env.cli_list_json().stdout_json();
-    assert_eq!(json[0]["state"], "Exited");
-    assert_eq!(json[0]["last_exit_code"], 0);
-}
-
-#[cfg(unix)]
-fn render_ddot_template(install_dir: &str, pid_dir: &str) -> Option<String> {
-    let tmpl_path = std::path::PathBuf::from(option_env!("DDOT_TEMPLATE_PATH")?);
-    let tmpl = std::fs::read_to_string(&tmpl_path)
-        .unwrap_or_else(|e| panic!("failed to read {}: {e}", tmpl_path.display()));
-    let rendered = tmpl
-        .replace("{{.InstallDir}}", install_dir)
-        .replace("{{.PIDDir}}", pid_dir);
-    Some(
-        rendered
-            .lines()
-            .filter(|line| !line.contains("{{"))
-            .collect::<Vec<_>>()
-            .join("\n"),
-    )
-}
-
-#[cfg(unix)]
-#[test]
-fn test_ddot_template_starts_with_env_and_optional_envfile() {
-    let dir = tempfile::tempdir().unwrap();
-
-    let install_dir = dir.path().join("install");
-    let bin_dir = install_dir.join("ext/ddot/embedded/bin");
-    std::fs::create_dir_all(&bin_dir).unwrap();
-
-    let conf_dir = dir.path().join("conf");
-    std::fs::create_dir_all(&conf_dir).unwrap();
-    let fleet_policies_dir = conf_dir.join("managed/datadog-agent/stable");
-
-    let script = bin_dir.join("otel-agent");
-    std::fs::write(
-        &script,
-        format!(
-            concat!(
-                "#!/bin/sh\n",
-                "if [ \"$DD_FLEET_POLICIES_DIR\" = \"{expected}\" ]; then\n",
-                "  exit 0\n",
-                "else\n",
-                "  echo \"DD_FLEET_POLICIES_DIR=$DD_FLEET_POLICIES_DIR\" >&2\n",
-                "  exit 1\n",
-                "fi\n",
-            ),
-            expected = fleet_policies_dir.to_str().unwrap(),
-        ),
-    )
-    .unwrap();
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        std::fs::set_permissions(&script, std::fs::Permissions::from_mode(0o755)).unwrap();
-    }
-
-    let pid_dir = dir.path().join("pid");
-    std::fs::create_dir_all(pid_dir.join("run")).unwrap();
-
-    let config_dir = dir.path().join("processes.d");
-    std::fs::create_dir_all(&config_dir).unwrap();
-
-    let Some(yaml) = render_ddot_template(install_dir.to_str().unwrap(), pid_dir.to_str().unwrap())
-    else {
-        eprintln!("DDOT_TEMPLATE_PATH not set at compile time, skipping");
-        return;
-    };
-    std::fs::write(config_dir.join("datadog-agent-ddot.yaml"), &yaml).unwrap();
-
-    let sock = dir.path().join("daemon.sock");
-    let mut daemon = helpers::DaemonHandle::start_with_env(
-        &config_dir,
-        &sock,
-        &[
-            ("DD_CONF_DIR", conf_dir.to_str().unwrap()),
-            ("DD_INSTALL_DIR", install_dir.to_str().unwrap()),
-            ("DD_PID_DIR", pid_dir.to_str().unwrap()),
-        ],
-    );
-    assert!(
-        daemon.wait_for_log_default(
-            "[datadog-agent-ddot] optional environment file not found, skipping"
-        ),
-        "daemon should skip the missing optional env file for ddot"
-    );
-    assert!(
-        daemon.wait_for_log_default("[datadog-agent-ddot] spawned (pid="),
-        "ddot process should be spawned with the otel-agent binary"
-    );
-    assert!(
-        daemon.wait_for_log_default("[datadog-agent-ddot] exited with exit status: 0"),
-        "ddot process should exit 0 (DD_FLEET_POLICIES_DIR env var was set correctly)"
-    );
-    assert!(
-        daemon.wait_for_log_default(
-            "[datadog-agent-ddot] exit does not match restart policy, not restarting"
-        ),
-        "on-failure restart should not trigger on exit 0"
-    );
-
-    let status = daemon.stop();
-    assert!(status.success());
-}
-
-#[cfg(unix)]
-#[test]
-fn test_ddot_template_skipped_when_binary_missing() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_dir = dir.path().join("processes.d");
-    std::fs::create_dir_all(&config_dir).unwrap();
-
-    let Some(yaml) = render_ddot_template("/nonexistent/install", "/nonexistent/pid") else {
-        eprintln!("DDOT_TEMPLATE_PATH not set at compile time, skipping");
-        return;
-    };
-    std::fs::write(config_dir.join("datadog-agent-ddot.yaml"), &yaml).unwrap();
-
-    let sock = dir.path().join("daemon.sock");
-    let mut daemon = helpers::DaemonHandle::start(&config_dir, &sock);
-    assert!(
-        daemon.wait_for_log_default("[datadog-agent-ddot] condition_path_exists not met"),
-        "daemon should skip ddot when otel-agent binary is missing"
-    );
-    assert_eq!(
-        daemon.count_log_matches("[datadog-agent-ddot] spawned"),
-        0,
-        "ddot should NOT be spawned when condition_path_exists is not met"
-    );
-
-    let status = daemon.stop();
-    assert!(status.success());
-}
