+++ pkg/procmgr/rust/tests/e2e/main.rs
@@ -1,0 +1,20 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2026-present Datadog, Inc.
+
+#[path = "../helpers/mod.rs"]
+mod helpers;
+
+mod catalog_list;
+mod cli_contracts;
+mod config_status;
+mod create;
+mod daemon;
+mod ddot;
+mod describe;
+mod env_files;
+mod fixtures;
+mod process_control;
+mod reload;
+mod restart_burst;
